Add fixtures for generating access-tokens
refs #116
