+++ project/access/management/commands/generate_tokens.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+from django.core.management.base import BaseCommand, CommandError
+from access.tests.fixtures.tokens import TokenFactory
+
+
+class Command(BaseCommand):
+    help = 'generate randomised (member) tokens'
+
+    def add_arguments(self, parser):
+        parser.add_argument('amount', type=int)
+
+    def handle(self, *args, **options):
+        for i in range(options['amount']):
+            token = TokenFactory()
+            print("Generated token %s" % token)
