+++ project/access/tests/fixtures/nonmembertokens.py
@@ -6,6 +6,8 @@ from asylum.utils import get_random_objects
 from asylum.tests.utils import FuzzyLoremipsum
 from members.tests.fixtures.memberlikes import lastnames, firstnames, generate_email
 from members.models import MemberCommon
+from .tokens import generate_value
+
 
 def generate_contact(x):
     tmp = MemberCommon()
@@ -13,14 +15,6 @@ def generate_contact(x):
     tmp.lname = random.choice(lastnames)
     tmp.email = generate_email(tmp)
     return "%s, %s <%s>" % (tmp.lname, tmp.fname, tmp.email)
-
-
-def generate_value(tokenlike):
-    if "DESFire" in tokenlike.ttype.label:
-        return "80" + "".join([ "%02x" % random.randint(1,255) for x in range(6) ])
-    if "Caller-id" in tokenlike.ttype.label:
-        return "555-" + "".join([ "%d" % random.randint(0,9) for x in range(8) ])
-    return "random %x" % random.randint(1,2**32)
 
 
 class NonMemberTokenFactory(factory.django.DjangoModelFactory):
